тест

+++ app/src/main/java/com/example/practic/Account.kt
@@ -2,7 +2,7 @@ package com.example.practic
 
 interface Account {
     fun getCurrencyName(): String
-    fun addMoney(money: Double)
-    fun getBalance(): Double
-    fun getMoney(money: Double): Double
+    fun addMoney(money: Int)
+    fun getBalance(): Int
+    fun getMoney(money: Int): Int
 }
