+++ app/src/main/java/com/example/practic/RubleAccaunt.kt
@@ -1,24 +1,24 @@
 package com.example.practic
 
 class RubleAccaunt : Account {
-    var ruble = 0.0
+    var ruble = 0
     override fun getCurrencyName(): String {
         return "Ruble"
     }
 
-    override fun addMoney(money: Double) {
+    override fun addMoney(money: Int) {
         if (money < 0) return
 
         ruble += money
     }
 
-    override fun getBalance(): Double {
+    override fun getBalance(): Int {
 
         return ruble
     }
 
-    override fun getMoney(money: Double): Double {
-        if (money > ruble) return 0.0
+    override fun getMoney(money: Int): Int {
+        if (money > ruble) return 0
         ruble -= money
         return ruble
     }
